beginning of tcgplayer api script. market_price_scraper refactor
tcgplayer recently adjusted their card/product html structure, resulting in the market_price_scraper not being able to decipher the difference between cards and products. This created a bpc product for every card, no new cards are being added. Instead of fixing it the plan is to substitute in tcgpapi script