added pagination to expansions and items
item and expansion response includes items and total count. Refactored searchValueToWhereLike to exclude where statement so it can be combined with additional where statements. Expansion and item endpoints take in optional params to filter based on searchvalue, expansionId, whether to include printings or not, also what attribute to sort by and which direction